refactor(helm): split helpers.py into role-named modules
helpers.py mixed client construction with the unavailable tool envelope. Split them so each file names its role.
