+++ tests/integrations/helm/test_client_factory.py
@@ -1,4 +1,4 @@
-"""Tests for shared Helm integration helpers."""
+"""Tests for Helm client construction from raw tool params."""
 
 from __future__ import annotations
 
@@ -8,7 +8,7 @@ import pytest
 from pydantic import ValidationError
 
 from integrations.config_models import HelmIntegrationConfig
-from integrations.helm import helpers
+from integrations.helm import client_factory
 from integrations.helm.client import HelmClient
 
 
@@ -19,7 +19,7 @@ def _helm_validation_error() -> ValidationError:
 
 
 def test_helm_client_for_run_builds_client_for_valid_config() -> None:
-    client = helpers.helm_client_for_run(integration_id="helm-test")
+    client = client_factory.helm_client_for_run(integration_id="helm-test")
 
     assert isinstance(client, HelmClient)
 
@@ -29,10 +29,10 @@ def test_helm_client_for_run_returns_none_for_validation_error(
 ) -> None:
     validate = MagicMock(side_effect=_helm_validation_error())
     debug_log = MagicMock()
-    monkeypatch.setattr(helpers.HelmIntegrationConfig, "model_validate", validate)
-    monkeypatch.setattr(helpers.logger, "debug", debug_log)
+    monkeypatch.setattr(client_factory.HelmIntegrationConfig, "model_validate", validate)
+    monkeypatch.setattr(client_factory.logger, "debug", debug_log)
 
-    assert helpers.helm_client_for_run() is None
+    assert client_factory.helm_client_for_run() is None
     debug_log.assert_not_called()
 
 
@@ -42,10 +42,10 @@ def test_helm_client_for_run_logs_unexpected_validation_failure(
     error = RuntimeError("unexpected validation failure")
     validate = MagicMock(side_effect=error)
     debug_log = MagicMock()
-    monkeypatch.setattr(helpers.HelmIntegrationConfig, "model_validate", validate)
-    monkeypatch.setattr(helpers.logger, "debug", debug_log)
+    monkeypatch.setattr(client_factory.HelmIntegrationConfig, "model_validate", validate)
+    monkeypatch.setattr(client_factory.logger, "debug", debug_log)
 
-    assert helpers.helm_client_for_run() is None
+    assert client_factory.helm_client_for_run() is None
     debug_log.assert_called_once_with(
         "helm_client_for_run failed unexpectedly",
         exc_info=True,
